+++ src/mame/drivers/nes_vt32_vt369.cpp
@@ -23,6 +23,7 @@
 
 #include "emu.h"
 #include "machine/nes_vt32_vt369_soc.h"
+#include "machine/nes_vt32_soc.h"
 
 class nes_vt3x_base_state : public driver_device
 {
